pyproject.toml
[tool.poetry]
name = "previsao-inadimplencia"
version = "0.1.0"
description = ""
authors = ["Felipe Lucas Roll <felipelroll@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
pypyodbc = "^1.3.6"
python-dotenv = "^1.0.1"
plotly-express = "^0.4.1"
nbformat = "^5.10.4"
kaleido = "0.1.0.post1"
scikit-learn = "^1.5.2"
imbalanced-learn = "^0.12.4"
xgboost = "^2.1.2"
joblib = "^1.4.2"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
